% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepG4.R
\name{DeepG4}
\alias{DeepG4}
\title{DeepG4 main function to predict a probability to form a G4, given a DNA sequence.}
\usage{
DeepG4(
  X = NULL,
  Y = NULL,
  model = NULL,
  tabv = c(N = 5, T = 4, G = 3, C = 2, A = 1),
  lower.case = F,
  seq.size = 201,
  treshold = 0.5
)
}
\arguments{
\item{X}{a character, a list or a DNAStringSet/DNAStringSetList of DNA sequences.}

\item{Y}{a numeric vector of 1 and 0 values (default to NULL).}

\item{model}{a path to a keras model in hdf5 format (default to NULL). Don't change it unless you want to use our function with a custom model.}

\item{tabv}{a named vector of numeric values representing the DNA to numerical conversion. Don't change it unless you want to use our function with a custom model.}

\item{lower.case}{boolean. Set to \code{TRUE} if elements of X are in lower case (default to FALSE).}

\item{seq.size}{numeric value representing the sequence size accepted by our model. Don't change it unless you want to use our function with a custom model.}

\item{treshold}{numeric value who define the treshold to use to get confusion matrix (default to 0.5).}
}
\value{
if \code{Y = NULL}, return DeepG4 prediction for each value of X.
    if \code{Y} is provided, return a list with list(prediction for each value of X,a ggplot2 object representing AUC,a ggplot2 object representing confusion matrix,some metrics)
}
\description{
DeepG4 main function to predict a probability to form a G4, given a DNA sequence.
}
\details{
This function is a wrapper to help people to get a prediction given any DNA sequence(s) of type ACGTN with our DeepG4 model.
 You don't have to use it to get a DeepG4 prediction, if you're familar with keras and tensorflow, you can access our model in hdf5 package using \code{system.file("extdata", "model.hdf5", package = "DeepG4")}.
 In complement, \code{\link{DNAToNumerical}} can help you to get the one-hot conversion needed by our model as input.
 If your sequences > \code{seq.size}, they will be cropped and sequences < \code{seq.size}, will be filled with zero padding.
}
\examples{
library(Biostrings)
library(DeepG4)

sequences <- system.file("extdata", "test_G4_data.fa", package = "DeepG4")
sequences <- readDNAStringSet(sequences)

predictions <- DeepG4(sequences)
head(predictions)
}
