% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepG4InputFromBED.R
\name{DeepG4InputFromBED}
\alias{DeepG4InputFromBED}
\title{Pre-Computing function of DeepG4. To use before DeepG4 main function}
\usage{
DeepG4InputFromBED(
  BED = NULL,
  ATAC = NULL,
  is.bw = TRUE,
  GENOME = NULL,
  use.bg = TRUE,
  windows_bg = 5000,
  treshold_bg = 2
)
}
\arguments{
\item{BED}{An object of class GRanges.}

\item{ATAC}{A character path of bigWig/bedGraph file, or an object of class GRanges/SimpleRleList.}

\item{is.bw}{a boolean. Set to \code{TRUE} if you want to use rtracklayer::import.bw fonction, \code{FALSE} use rtracklayer::import.bedGraph.}

\item{GENOME}{a BSgenome object containing  the DNA sequence of genome of interest.}

\item{use.bg}{a boolean. Set to \code{TRUE} you want to normalize the accessibility using a windows background of windows_bg.}

\item{windows_bg}{numeric value who define the windows use to get background signal.}

\item{treshold_bg}{numeric value who set the treshold signal/background.}
}
\value{
a list(DNAStringSet,vector) with fasta sequence from GENOME of BED, and vector of accessibility of BED.
}
\description{
Pre-Computing function of DeepG4. To use before DeepG4 main function
}
