% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepG4Scan.R
\name{DeepG4Scan}
\alias{DeepG4Scan}
\title{Standard method to call}
\usage{
DeepG4Scan(X, ...)
}
\arguments{
\item{X}{One of character, list, DNAString, DNAStringSet, DNAStringSetList, GRanges}

\item{...}{}
}
\value{

}
\description{
Standard method to call
}
\details{
This function is a method who launch ?callDeepG4Scan or .callDeepG4ScanATAC based on the arguments who are passed on. See ?.callDeepG4ScanATAC or ?.callDeepG4Scan.
}
