% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAToNumerical.R
\name{DNAToNumerical}
\alias{DNAToNumerical}
\title{Convert DNAStringSet object into one-hot encoding}
\usage{
DNAToNumerical(
  x,
  tabv = c(N = 5, T = 4, G = 3, C = 2, A = 1),
  lower.case = F,
  seq.size = 201
)
}
\arguments{
\item{x}{DNAStringSet object}

\item{tabv}{named vector list of numerical values which indicate the numerical value of a nucleotide.}

\item{lower.case}{Set to \code{TRUE} if the DNA in your sequence is encoded in lower.case i.e. "acgt", default set to \code{FALSE}.}

\item{seq.size}{Set the sequence maximal length value authorized by our model (default to 201).}
}
\value{
An array of dimension \code{nrow(x),ncol(x),length(tabv)}
}
\description{
Convert DNAStringSet object into one-hot encoding
}
\examples{
x <- Biostrings::DNAStringSet(c("ACGT"))
x_onehot <- DNAToNumerical(x)
x_onehot
}
