% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepG4Scan.R
\name{.callDeepG4ScanATAC}
\alias{.callDeepG4ScanATAC}
\title{Scanning of potential active G4 on a sequence or a list of sequences using a sliding window of size k.}
\usage{
.callDeepG4ScanATAC(
  X = NULL,
  X.ATAC = NULL,
  k = 20,
  treshold = 0.5,
  threads = 1,
  lower.case = F,
  GENOME = NULL,
  use.bg = T,
  windows_bg = 5000,
  treshold_bg = 2
)
}
\arguments{
\item{X}{An object of class GRanges.}

\item{X.ATAC}{An object of class GRanges containing DNA accessibility.}

\item{k}{size of the sliding windows.}

\item{treshold}{numeric value who define the treshold to use to consider a sequence asc ontaining an active G4.}

\item{threads}{numeric value who define the number of threads used in DeepG4Scan (Generate sub sequences)}

\item{lower.case}{a boolean. Set to \code{TRUE} if elements of X are in lower case (default to FALSE).}

\item{GENOME}{a BSgenome object containing the DNA sequence of genome of interest.}

\item{use.bg}{a boolean. Set to \code{TRUE} you want to normalize the accessibility using a windows background of windows_bg.}

\item{windows_bg}{numeric value who define the windows use to get background signal.}

\item{treshold_bg}{numeric value who set the treshold signal/background.}
}
\value{
a data.frame with the position of potential active G4 across input sequences.
}
\description{
Scanning of potential active G4 on a sequence or a list of sequences using a sliding window of size k.
}
