% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractMotifFromModel.R
\name{ExtractMotifFromModel}
\alias{ExtractMotifFromModel}
\title{ExtractMotifFromModel allow the user to extract motifs who are activated by the sequences in our model.}
\usage{
ExtractMotifFromModel(X = NULL, Y = NULL, lower.case = F, top_kernel = 20)
}
\arguments{
\item{X}{An object of class character,list or DNAStringSet/DNAStringSetList with DNA sequences.}

\item{Y}{a numeric vector of 1 and 0 values (default to NULL).}

\item{lower.case}{boolean. Set to \code{TRUE} if elements of X are in lower case (default to FALSE).}

\item{top_kernel}{integer which indicate the top best motifs activated by X.}
}
\value{
A list of Position Count Matrix.
}
\description{
ExtractMotifFromModel allow the user to extract motifs who are activated by the sequences in our model.
}
\details{
A kernel is a weight matrix used in convolution layer in DNN (Deep Neural Network).
With one-hot DNA, a kernel represented a weighted DNA motif, but can't be easily interpreted.
With this function, we detect the location on the response map where the kernel is activated for all input sequences and
 built the corresponding position count matrix.
}
