% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractSubSequencefromBed.R
\name{ExtractSubSequencefromBed}
\alias{ExtractSubSequencefromBed}
\title{ExtractSubSequencefromBed : Internal function for DeepG4Scan.}
\usage{
ExtractSubSequencefromBed(
  x = NULL,
  x.atac = NULL,
  k = 20,
  GENOME = NULL,
  seq.size = 201,
  nb.threads = 1,
  use.bg = T,
  windows_bg = 5000,
  treshold_bg = 2
)
}
\arguments{
\item{x}{An object of class GRanges.}

\item{x.atac}{A SimpleRleList object containing DNA accessibility.}

\item{k}{size of the sliding windows.}

\item{GENOME}{a BSgenome object containing  the DNA sequence of genome of interest.}

\item{seq.size}{numeric value representing the sequence size accepted by our model. Don't change it unless you want to use our function with a custom model.}

\item{nb.threads}{number of threads to use, default 1.}

\item{use.bg}{a boolean. Set to \code{TRUE} you want to normalize the accessibility using a windows background of windows_bg.}

\item{windows_bg}{numeric value who define the windows use to get background signal.}

\item{treshold_bg}{numeric value who set the treshold signal/background.}
}
\value{
A data.frame with position of subsequence, DNA sequence and accessibility.
}
\description{
ExtractSubSequencefromBed : Internal function for DeepG4Scan.
}
