% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScoreBW.R
\name{getScoreBW}
\alias{getScoreBW}
\title{getScoreBW extract DNA accessibility from WIG to BED positions.}
\usage{
getScoreBW(WIG, BED, forScan = F)
}
\arguments{
\item{WIG}{An object of class GRanges containing DNA accessibility.}

\item{BED}{An object of class GRanges.}
}
\value{
A SimpleRleList object containing DNA accessibility.
}
\description{
getScoreBW extract DNA accessibility from WIG to BED positions.
}
