% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormBW.R
\name{NormBW}
\alias{NormBW}
\title{Normalize a coverage file (from bigWig or bedGraph) using scales::rescale from a specific set of ranges to set values between [0,1]}
\usage{
NormBW(x, binbed)
}
\arguments{
\item{x}{An object of class GRanges, the coverage file to be normalized}

\item{binbed}{An object of class GRanges, the genomic position used to normalize x.}
}
\value{

}
\description{
Normalize a coverage file (from bigWig or bedGraph) using scales::rescale from a specific set of ranges to set values between [0,1]
}
